添加了问卷列表查询， 修改获取题目列表

+++ hp-parent/hp-wx-mobile/src/main/java/com/hp/mobile/mapper/UserAnswerMapper.java
@@ -10,7 +10,7 @@ public interface UserAnswerMapper {
 
     int insert(UserAnswer record);
 
-    int insertSelective(UserAnswer record);
+    Long insertSelective(UserAnswer record);
 
     UserAnswer selectByPrimaryKey(Long objectId);
